backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "tepennnn",email = "stephen.nathaniel.handiyapna@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.127.0,<0.128.0)",
    "fastapi-cli (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
